pyproject.toml
[project]
name = "decap_correction_tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.36.2",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "streamlit>=1.46.1",
]
